DL/requirements.txt
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
cloudpickle==2.2.0
contourpy==1.0.5
cycler==0.11.0
dotmap==1.3.30
fonttools==4.37.4
gym==0.23.1
gym-notices==0.0.8
idna==3.4
iniconfig==1.1.1
kiwisolver==1.4.4
matplotlib==3.6.1
numpy==1.23.4
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pybullet==3.2.5
pygame==2.1.2
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
six==1.16.0
tomli==2.0.1
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
typing_extensions==4.4.0
urllib3==1.26.12
